+++ src/operators.jl
@@ -262,6 +262,200 @@ function Base.dot{N}(lhs::JuMPArray{Float64,N},rhs::JuMPArray{Float64,N})
     return sum(lhs.innerArray .* rhs.innerArray)
 end
 
+Base.promote_rule{R<:Real}(::Type{Variable},::Type{R}       ) = AffExpr
+Base.promote_rule         (::Type{Variable},::Type{AffExpr} ) = AffExpr
+Base.promote_rule         (::Type{Variable},::Type{QuadExpr}) = QuadExpr
+Base.promote_rule{R<:Real}(::Type{AffExpr}, ::Type{R}       ) = AffExpr
+Base.promote_rule         (::Type{AffExpr}, ::Type{QuadExpr}) = QuadExpr
+Base.promote_rule{R<:Real}(::Type{QuadExpr},::Type{R}       ) = QuadExpr
+_throw_transpose_error() = error("Transpose not currently implemented for JuMPArrays with arbitrary index sets.")
+
+Base.transpose(x::OneIndexedArray)  = transpose(x.innerArray)
+Base.transpose(x::JuMPArray)  = _throw_transpose_error()
+Base.ctranspose(x::OneIndexedArray) = ctranspose(x.innerArray)
+Base.ctranspose(x::JuMPArray)  = _throw_transpose_error()
+
+###############
+# The _multiply!(buf,y,z) adds the results of y*z into the buffer buf. No bounds/size
+# checks are performed; it is expected that the caller has done this, has ensured
+# that the eltype of buf is appropriate, and has zeroed the elements of buf (if desired).
+
+function _multiply!(ret, A, x)
+    m, n = size(A,1), size(A,2)
+    for i in 1:m, j in 1:n
+        ret[i] += A[i,j] * x[j]
+    end
+    return ret
+end
+
+_sizehint_expr!(q::AffExpr, n::Int) = begin
+        sizehint!(q.vars, n)
+        sizehint!(q.coeffs, n)
+end
+
+_sizehint_expr!(q::QuadExpr, n::Int) = begin
+        sizehint!(q.qvars1, n)
+        sizehint!(q.qvars2, n)
+        sizehint!(q.qcoeffs, n)
+        _sizehint_expr!(q.aff, n)
+end
+_sizehint_expr!(q, n) = q # for type stability
+
+function _multiply!{T<:Union(GenericAffExpr,GenericQuadExpr)}(ret::Array{T}, lhs, rhs)
+    m, n = size(lhs,1), size(lhs,2)
+    r, s = size(rhs,1), size(rhs,2)
+    p, q = size(ret,1), size(ret,2)
+    for i in 1:m, j in 1:s
+        q = T()
+        _sizehint_expr!(q, n)
+        for k in 1:n
+            tmp = convert(T, lhs[i,k]*rhs[k,j])
+            append!(q, tmp)
+        end
+        ret[i,j] = q
+    end
+    ret
+end
+
+function _multiply!{T<:Union(GenericAffExpr,GenericQuadExpr)}(ret::Array{T}, lhs::SparseMatrixCSC, rhs)
+    m, n = size(lhs,1), size(lhs,2)
+    r, s = size(rhs,1), size(rhs,2)
+    p, q = size(ret,1), size(ret,2)
+    for i in 1:m, j in 1:s
+        q = T()
+        _sizehint_expr!(q, n)
+        for k in lhs.colptr[i]:lhs.colptr[i+1]
+            tmp = convert(T, lhs.nzval[k]*rhs[lhs.rowval[k],j])
+            append!(q, tmp)
+        end
+        ret[i,j] = q
+    end
+    ret
+end
+
+# Not sure how/why you'd make a sparse matrix with Variables...
+_multiply!{T<:Union(GenericAffExpr,GenericQuadExpr)}(ret::Array{T}, lhs::SparseMatrixCSC, rhs::SparseMatrixCSC) =
+    _multiply!(ret, lhs, full(rhs))
+
+# Kind of hacky that this relies on commutativity...
+function _multiply!{T<:Union(GenericAffExpr,GenericQuadExpr)}(ret::Array{T}, lhs, rhs::SparseMatrixCSC)
+    tmp = Array(T, size(ret,2), size(ret,1))
+    _multiply!(tmp, rhs', lhs')
+    copy!(ret, tmp')
+    ret
+end
+
+typealias JuMPTypes Union(Variable,AffExpr,QuadExpr)
+
+(*)(lhs::AbstractArray, rhs::OneIndexedArray) = (*)(lhs, rhs.innerArray)
+(*)(lhs::OneIndexedArray, rhs::AbstractArray) = (*)(lhs.innerArray, rhs)
+(*)(lhs::OneIndexedArray, rhs::OneIndexedArray) = (*)(lhs.innerArray, rhs.innerArray)
+
+# Don't do size checks here in _return_array, defer that to (*)
+function _return_array{R,S}(A::AbstractArray{R}, x::AbstractArray{S,1})
+    Q = (R <: JuMPTypes && S <: JuMPTypes) ? QuadExpr : AffExpr
+    m = size(A,1)
+    Array(Q, m)
+end
+
+function _return_array{R,S}(A::AbstractArray{R}, x::AbstractArray{S,2})
+    Q = (R <: JuMPTypes && S <: JuMPTypes) ? QuadExpr : AffExpr
+    m = size(A,1)
+    s = size(x,2)
+    Array(Q, m, s)
+end
+
+function (*){T<:JuMPTypes}(A::Array{T}, x::Union(Array,SparseMatrixCSC))
+    m, n = size(A,1), size(A,2)
+    r, s = size(x,1), size(x,2)
+    n == r || error("Incompatible sizes")
+    ret = _return_array(A, x)
+    _multiply!(ret, A, x)
+    ret
+end
+
+function (*){T<:JuMPTypes,R<:JuMPTypes}(A::Array{T}, x::Array{R})
+    (T == QuadExpr || R == QuadExpr) && error("Cannot multiply two arrays of QuadExpr")
+    m, n = size(A,1), size(A,2)
+    r, s = size(x,1), size(x,2)
+    n == r || error("Incompatible sizes")
+    ret = _return_array(A, x)
+    _multiply!(ret, A, x)
+    ret
+end
+
+function (*){T<:JuMPTypes}(A::Union(Array,SparseMatrixCSC), x::Array{T})
+    m, n = size(A,1), size(A,2)
+    r, s = size(x,1), size(x,2)
+    n == r || error("Incompatible sizes")
+    ret = _return_array(A, x)
+    _multiply!(ret, A, x)
+    return ret
+end
+
+for op in [:+, :-, :*]
+    @eval begin
+        $op{T<:JuMPTypes}(lhs::Real,rhs::Array{T}) = map(c->$op(lhs,c), rhs)
+        $op{T<:JuMPTypes}(lhs::Array{T},rhs::Real) = map(c->$op(c,rhs), lhs)
+        $op(lhs::Real,rhs::OneIndexedArray) = $op(lhs, rhs.innerArray)
+        $op(lhs::OneIndexedArray,rhs::Real) = $op(lhs.innerArray, rhs)
+
+        $op{T<:JuMPTypes}(lhs::Array{T},rhs::OneIndexedArray) = $op(lhs,rhs.innerArray)
+        $op{T<:JuMPTypes}(lhs::OneIndexedArray,rhs::Array{T}) = $op(lhs.innerArray,rhs)
+        $op(lhs::OneIndexedArray,rhs::OneIndexedArray) = $op(lhs.innerArray,rhs.innerArray)
+    end
+end
+
+# The following are primarily there for internal use in the macro code for @addConstraint
+for op in [:(+), :(-)]; @eval begin
+    function $op(lhs::GenericAffExpr, rhs::Array)
+        (isempty(lhs.vars) && isempty(lhs.coeffs)) || error("Cannot perform $typeof(lhs)) + $(typeof(rhs))")
+        $op(lhs.constant, rhs)
+    end
+    function $op(lhs::Array, rhs::GenericAffExpr)
+        (isempty(rhs.vars) && isempty(rhs.coeffs)) || error("Cannot perform $typeof(lhs)) + $(typeof(rhs))")
+        $op(lhs, rhs.constant)
+    end
+    function $op(lhs::GenericQuadExpr, rhs::Array)
+        (isempty(lhs.qvars1) && isempty(lhs.qvars2) && isempty(lhs.qcoeffs) &&
+            isempty(lhs.aff.vars) && isempty(lhs.aff.coeffs)) ||
+            error("Cannot perform $typeof(lhs)) + $(typeof(rhs))")
+        $op(lhs.aff.constant, rhs)
+    end
+    function $op(lhs::Array, rhs::GenericQuadExpr)
+        (isempty(rhs.qvars1) && isempty(rhs.qvars2) && isempty(rhs.qcoeffs) &&
+            isempty(rhs.aff.vars) && isempty(rhs.aff.coeffs)) ||
+            error("Cannot perform $typeof(lhs)) + $(typeof(rhs))")
+        $op(lhs, rhs.aff.constant)
+    end
+end; end
+
+(/){T<:JuMPTypes}(lhs::Array{T},rhs::Real) = map(c->$op(c,rhs), lhs)
+(/)(lhs::OneIndexedArray,rhs::Real) = $op(lhs.innerArray, rhs)
+
+for (dotop,op) in [(:.+,:+), (:.-,:-), (:.*,:*), (:./,:/)]
+    @eval begin
+        $dotop(lhs::Real,rhs::JuMPTypes) = $op(lhs,rhs)
+        $dotop(lhs::JuMPTypes,rhs::Real) = $op(lhs,rhs)
+        $dotop{T<:JuMPTypes}(lhs::Real,rhs::Array{T}) = map(c->$op(lhs,c), rhs)
+        $dotop{T<:JuMPTypes}(lhs::Array{T},rhs::Real) = map(c->$op(c,rhs), lhs)
+        $dotop(lhs::Real,rhs::OneIndexedArray) = $dotop(lhs, rhs.innerArray)
+        $dotop(lhs::OneIndexedArray,rhs::Real) = $dotop(lhs.innerArray, rhs)
+        $dotop{T<:JuMPTypes,S<:JuMPTypes}(lhs::T,rhs::Array{S}) = map(c->$op(lhs,c), rhs)
+        $dotop{T<:JuMPTypes,S<:JuMPTypes}(lhs::Array{T},rhs::S) = map(c->$op(c,rhs), lhs)
+
+        $dotop{T<:JuMPTypes,N}(lhs::Array{T,N},rhs::OneIndexedArray) = $dotop(lhs,rhs.innerArray)
+        $dotop{T<:JuMPTypes,N}(lhs::OneIndexedArray,rhs::Array{T,N}) = $dotop(lhs.innerArray,rhs)
+        $dotop(lhs::OneIndexedArray,rhs::OneIndexedArray) = $dotop(lhs.innerArray,rhs.innerArray)
+    end
+end
+(+){T<:JuMPTypes}(x::Array{T}) = x
+(+)(x::OneIndexedArray) = x.innerArray
+(-)(x::Array{Variable}) = (-)(convert(Array{AffExpr},x))
+(-)(x::OneIndexedArray) = -x.innerArray
+(*){T<:JuMPTypes}(x::Array{T}) = x
+(*)(x::OneIndexedArray) = x.innerArray
+
 ###############################################################################
 # Add nonlinear function fallbacks for JuMP built-in types
 const op_hint = "Are you trying to build a nonlinear problem? Make sure you use @addNLConstraint/@setNLObjective."